Revamp aliases, and much more on the way to codegen
Many of the hdb views and their INSTEAD OF triggers look very much
alike.  The one that stood out was hdb.aliases, which was very different
due to how we were modelling aliases in the heimdal schema.  By
re-thinking that, the hdb.aliases view and its triggers are now a lot
more like the others, which means we could soon take a representative
view and triggers and templatize them so that all the views and triggers
other than the triggers on hdb.hdb can be auto-generated.

A bunch other cleanup was done as well, such as more consistently using
underscores for identifiers in JSON object keys.